Upgrade to fontawesome 5
Shifting from the external CND's v4.7.0 fontawesome to the local git submodule
(which is version 5). I'm new to it but from what I read the main difference
looks to be that we should use 'fas' to indicate the solid icon family. This
doesn't seem to be necessary (importing the solid stylesheet and keeping just
'fa' does the same) but this seems to be its more proper use.