Wrap contact form in a div
I found this W3C validation error kinda odd...

  Error Line 61, Column 60: document type does not allow element "INPUT" here;
  missing one of "P", "H1", "H2", "H3", "H4", "H5", "H6", "PRE", "DIV",
  "ADDRESS" start-tag

Evidently form tags are expected to only contain block elements. Many examples
(including W3C's own for form tags) don't do this. A better fix might be to
define ourselves as html5 in the header, but that has a higher risk of
compatibility issues. Simply wrapping this in a div does the trick and doesn't
look to have a visible impact so taking the more minimal approach.

  https://stackoverflow.com/questions/5865963/w3c-validator-gives-document-type-does-not-allow-element-input-here-error