Implement email form
Wiring up our contact form. The php is effectively a copy-paste of my old
site's. Unlike my old site I'm echoing the message back if it fails to be
sent (losing your text when there's a bug is infuriating).

I'm not entirely sure yet if this works. Locally it fails, but that might
simply be related to my local apache configuration.