Use PHPMailer
Migrating from php's builtin mail() function to PHPMailer. This is based
upon...

  https://github.com/PHPMailer/PHPMailer/blob/master/examples/sendmail.phps

My webserver was previously using sendmail, so I cannot exercise this until
it's located there. As such this is *not* tested and may require reversion or
adjustments.

Luckily this is my personal repo anyway, so it's all good. :P